+++ packages/connectors/src/coinbaseWallet.ts
@@ -1,7 +1,7 @@
 import type {
-  CoinbaseWalletSDK,
   Preference,
   ProviderInterface,
+  createCoinbaseWalletSDK,
 } from '@coinbase/wallet-sdk'
 import {
   ChainNotConfiguredError,
@@ -62,14 +62,16 @@ export function coinbaseWallet<version extends Version>(
 
 type Version4Parameters = Mutable<
   Omit<
-    ConstructorParameters<typeof CoinbaseWalletSDK>[0],
-    'appChainIds' // set via wagmi config
+    Parameters<typeof createCoinbaseWalletSDK>[0],
+    | 'appChainIds' // set via wagmi config
+    | 'preference'
   > & {
+    // TODO(v3): Remove `Preference['options']`
     /**
      * Preference for the type of wallet to display.
      * @default 'all'
      */
-    preference?: Preference['options'] | undefined
+    preference?: Preference['options'] | Compute<Preference> | undefined
   }
 >
 
@@ -79,7 +81,6 @@ function version4(parameters: Version4Parameters) {
     close?(): void
   }
 
-  let sdk: CoinbaseWalletSDK | undefined
   let walletProvider: Provider | undefined
 
   let accountsChanged: Connector['onAccountsChanged'] | undefined
@@ -171,24 +172,23 @@ function version4(parameters: Version4Parameters) {
     },
     async getProvider() {
       if (!walletProvider) {
-        // Unwrapping import for Vite compatibility.
-        // See: https://github.com/vitejs/vite/issues/9703
-        const CoinbaseWalletSDK = await (async () => {
-          const { default: SDK } = await import('@coinbase/wallet-sdk')
-          if (typeof SDK !== 'function' && typeof SDK.default === 'function')
-            return SDK.default
-          return SDK as unknown as typeof SDK.default
+        const preference = (() => {
+          if (typeof parameters.preference === 'string')
+            return { options: parameters.preference }
+          return {
+            ...parameters.preference,
+            options: parameters.preference?.options ?? 'all',
+          }
         })()
 
-        sdk = new CoinbaseWalletSDK({
+        const { createCoinbaseWalletSDK } = await import('@coinbase/wallet-sdk')
+        const sdk = createCoinbaseWalletSDK({
           ...parameters,
           appChainIds: config.chains.map((x) => x.id),
+          preference,
         })
 
-        walletProvider = sdk.makeWeb3Provider({
-          ...parameters,
-          options: parameters.preference ?? 'all',
-        })
+        walletProvider = sdk.getProvider()
       }
 
       return walletProvider