+++ packages/core/src/actions/getBalance.test.ts
@@ -1,6 +1,6 @@
 import { accounts, chain, config, testClient } from '@wagmi/test'
 import { parseEther } from 'viem'
-import { beforeEach, describe, expect, test } from 'vitest'
+import { beforeEach, expect, test } from 'vitest'
 
 import { getBalance } from './getBalance.js'
 
@@ -19,11 +19,8 @@ beforeEach(async () => {
   await testClient.mainnet2.mine({ blocks: 1 })
 })
 
-describe('getBalance', () => {
-  test('default', async () => {
-    await expect(
-      getBalance(config, { address }),
-    ).resolves.toMatchInlineSnapshot(`
+test('default', async () => {
+  await expect(getBalance(config, { address })).resolves.toMatchInlineSnapshot(`
       {
         "decimals": 18,
         "formatted": "10000",
@@ -32,13 +29,11 @@ describe('getBalance', () => {
       }
     `)
 
-    await testClient.mainnet.setBalance({
-      address,
-      value: parseEther('6969.12222215666'),
-    })
-    await expect(
-      getBalance(config, { address }),
-    ).resolves.toMatchInlineSnapshot(`
+  await testClient.mainnet.setBalance({
+    address,
+    value: parseEther('6969.12222215666'),
+  })
+  await expect(getBalance(config, { address })).resolves.toMatchInlineSnapshot(`
       {
         "decimals": 18,
         "formatted": "6969.12222215666",
@@ -46,12 +41,12 @@ describe('getBalance', () => {
         "value": 6969122222156660000000n,
       }
     `)
-  })
+})
 
-  test('parameters: chainId', async () => {
-    await expect(
-      getBalance(config, { address, chainId: chain.mainnet2.id }),
-    ).resolves.toMatchInlineSnapshot(`
+test('parameters: chainId', async () => {
+  await expect(
+    getBalance(config, { address, chainId: chain.mainnet2.id }),
+  ).resolves.toMatchInlineSnapshot(`
       {
         "decimals": 18,
         "formatted": "420",
@@ -59,15 +54,15 @@ describe('getBalance', () => {
         "value": 420000000000000000000n,
       }
     `)
-  })
+})
 
-  test('parameters: token', async () => {
-    await expect(
-      getBalance(config, {
-        address: '0x4557B18E779944BFE9d78A672452331C186a9f48',
-        token: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
-      }),
-    ).resolves.toMatchInlineSnapshot(`
+test('parameters: token', async () => {
+  await expect(
+    getBalance(config, {
+      address: '0x4557B18E779944BFE9d78A672452331C186a9f48',
+      token: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
+    }),
+  ).resolves.toMatchInlineSnapshot(`
       {
         "decimals": 18,
         "formatted": "0.559062564299199392",
@@ -75,12 +70,28 @@ describe('getBalance', () => {
         "value": 559062564299199392n,
       }
     `)
-  })
+})
 
-  test('parameters: unit', async () => {
-    await expect(
-      getBalance(config, { address, unit: 'wei' }),
-    ).resolves.toMatchInlineSnapshot(`
+test('parameters: token (bytes32 symbol)', async () => {
+  await expect(
+    getBalance(config, {
+      address: '0x4557B18E779944BFE9d78A672452331C186a9f48',
+      token: '0x9f8f72aa9304c8b593d555f12ef6589cc3a579a2',
+    }),
+  ).resolves.toMatchInlineSnapshot(`
+    {
+      "decimals": 18,
+      "formatted": "0",
+      "symbol": "MKR",
+      "value": 0n,
+    }
+  `)
+})
+
+test('parameters: unit', async () => {
+  await expect(
+    getBalance(config, { address, unit: 'wei' }),
+  ).resolves.toMatchInlineSnapshot(`
       {
         "decimals": 18,
         "formatted": "10000000000000000000000",
@@ -88,5 +99,4 @@ describe('getBalance', () => {
         "value": 10000000000000000000000n,
       }
     `)
-  })
 })