+++ site/.vitepress/theme/composables/useSponsors.ts
@@ -34,6 +34,11 @@ export function useSponsors() {
           url: 'https://paradigm.xyz',
           img: 'paradigm-light.svg',
         },
+        {
+          name: 'Ithaca',
+          url: 'https://ithaca.xyz',
+          img: 'ithaca-light.svg',
+        },
       ],
       gold: [
         {
@@ -152,6 +157,16 @@ export function useSponsors() {
           name: 'Thirdweb',
           url: 'https://thirdweb.com',
           img: 'thirdweb-light.svg',
+        },
+        {
+          name: 'Uxuy',
+          url: 'https://uxuy.com',
+          img: 'uxuy-light.svg',
+        },
+        {
+          name: 'Polymarket',
+          url: 'https://polymarket.com',
+          img: 'polymarket-light.svg',
         },
       ],
     }